+++ example/app/serializers.py
@@ -1,6 +1,8 @@
+from app.models import User
 from pydantic import BaseModel, constr
 
 from panther.file_handler import File, Image
+from panther.serializer import ModelSerializer
 
 
 class UserInputSerializer(BaseModel):
@@ -12,8 +14,9 @@ class UserOutputSerializer(BaseModel):
     username: str
 
 
-class UserUpdateSerializer(BaseModel):
-    username: str
+class UserUpdateSerializer(metaclass=ModelSerializer, model=User):
+    fields = ['username']
+    required_fields = ['username']
 
 
 class FileSerializer(BaseModel):