+++ panther/base_websocket.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 import asyncio
 import contextlib
 import logging
-from typing import TYPE_CHECKING
+from multiprocessing import Manager
+from typing import TYPE_CHECKING, Literal
 
 import orjson as json
 
@@ -11,19 +12,35 @@ from panther import status
 from panther._utils import generate_ws_connection_id
 from panther.base_request import BaseRequest
 from panther.configs import config
+from panther.db.connection import redis
 from panther.utils import Singleton
 
 if TYPE_CHECKING:
     from redis import Redis
 
-
 logger = logging.getLogger('panther')
 
 
+class PubSub:
+    def __init__(self, manager):
+        self._manager = manager
+        self._subscribers = self._manager.list()
+
+    def subscribe(self):
+        queue = self._manager.Queue()
+        self._subscribers.append(queue)
+        return queue
+
+    def publish(self, msg):
+        for queue in self._subscribers:
+            queue.put(msg)
+
+
 class WebsocketConnections(Singleton):
-    def __init__(self):
+    def __init__(self, manager: Manager = None):
         self.connections = {}
         self.connections_count = 0
+        self.manager = manager
 
     def __call__(self, r: Redis | None):
         if r:
@@ -31,39 +48,61 @@ class WebsocketConnections(Singleton):
             subscriber.subscribe('websocket_connections')
             logger.info("Subscribed to 'websocket_connections' channel")
             for channel_data in subscriber.listen():
-                # Check Type of PubSub Message
                 match channel_data['type']:
+                    # Subscribed
                     case 'subscribe':
                         continue
 
+                    # Message Received
                     case 'message':
                         loaded_data = json.loads(channel_data['data'].decode())
-                        if (
-                                isinstance(loaded_data, dict)
-                                and (connection_id := loaded_data.get('connection_id'))
-                                and (data := loaded_data.get('data'))
-                                and (action := loaded_data.get('action'))
-                                and (connection := self.connections.get(connection_id))
-                       ):
-                            # Check Action of WS
-                            match action:
-                                case 'send':
-                                    logger.debug(f'Sending Message to {connection_id}')
-                                    asyncio.run(connection.send(data=data))
-                                case 'close':
-                                    with contextlib.suppress(RuntimeError):
-                                        asyncio.run(connection.close(code=data['code'], reason=data['reason']))
-                                        # We are trying to disconnect the connection between a thread and a user
-                                        # from another thread, it's working, but we have to find another solution it
-                                        #
-                                        # Error:
-                                        # Task <Task pending coro=<Websocket.close()>> got Future
-                                        # <Task pending coro=<WebSocketCommonProtocol.transfer_data()>>
-                                        # attached to a different loop
-                                case _:
-                                    logger.debug(f'Unknown Message Action: {action}')
-                    case _:
-                        logger.debug(f'Unknown Channel Type: {channel_data["type"]}')
+                        self._handle_received_message(received_message=loaded_data)
+
+                    case unknown_type:
+                        logger.debug(f'Unknown Channel Type: {unknown_type}')
+        else:
+            self.pubsub = PubSub(manager=self.manager)
+            queue = self.pubsub.subscribe()
+            logger.info("Subscribed to 'websocket_connections' queue")
+            while True:
+                received_message = queue.get()
+                self._handle_received_message(received_message=received_message)
+
+    def _handle_received_message(self, received_message):
+        if (
+                isinstance(received_message, dict)
+                and (connection_id := received_message.get('connection_id'))
+                and connection_id in self.connections
+                and 'action' in received_message
+                and 'data' in received_message
+        ):
+            # Check Action of WS
+            match received_message['action']:
+                case 'send':
+                    asyncio.run(self.connections[connection_id].send(data=received_message['data']))
+                case 'close':
+                    with contextlib.suppress(RuntimeError):
+                        asyncio.run(self.connections[connection_id].close(
+                            code=received_message['data']['code'],
+                            reason=received_message['data']['reason']
+                        ))
+                        # We are trying to disconnect the connection between a thread and a user
+                        # from another thread, it's working, but we have to find another solution for it
+                        #
+                        # Error:
+                        # Task <Task pending coro=<Websocket.close()>> got Future
+                        # <Task pending coro=<WebSocketCommonProtocol.transfer_data()>>
+                        # attached to a different loop
+                case unknown_action:
+                    logger.debug(f'Unknown Message Action: {unknown_action}')
+
+    def publish(self, connection_id: str, action: Literal['send', 'close'], data: any):
+        publish_data = {'connection_id': connection_id, 'action': action, 'data': data}
+
+        if redis.is_connected:
+            redis.publish('websocket_connections', json.dumps(publish_data))
+        else:
+            self.pubsub.publish(publish_data)
 
     async def new_connection(self, connection: Websocket) -> None:
         await connection.connect(**connection.path_variables)
@@ -106,6 +145,7 @@ class Websocket(BaseRequest):
         pass
 
     async def send(self, data: any = None) -> None:
+        logger.debug(f'Sending WS Message to {self.connection_id}')
         if data:
             if isinstance(data, bytes):
                 await self.send_bytes(bytes_data=data)
@@ -121,6 +161,7 @@ class Websocket(BaseRequest):
         await self.asgi_send({'type': 'websocket.send', 'bytes': bytes_data})
 
     async def close(self, code: int = status.WS_1000_NORMAL_CLOSURE, reason: str = '') -> None:
+        logger.debug(f'Closing WS Connection {self.connection_id}')
         self.is_connected = False
         config['websocket_connections'].remove_connection(self)
         await self.asgi_send({'type': 'websocket.close', 'code': code, 'reason': reason})